+++ frontend/src/components/MyStats/RecentSessions/SessionsBox/index.tsx
