+++ frontend/src/components/MyStats/Stats/StatsCard/index.tsx
